+++ features/fixtures/ios-swift-cocoapods/iOSTestApp/scenarios/AutoSessionUnhandledScenario.h
@@ -1,0 +1,13 @@
+//
+//  AutoSessionUnhandledScenario.h
+//  iOSTestApp
+//
+//  Created by Delisa on 7/13/18.
+//  Copyright © 2018 Bugsnag. All rights reserved.
+//
+
+#import "Scenario.h"
+
+@interface AutoSessionUnhandledScenario : Scenario
+
+@end