+++ features/fixtures/ios-swift-cocoapods/iOSTestApp/scenarios/Scenario.h
@@ -6,10 +6,6 @@
 #import <Foundation/Foundation.h>
 #import <Bugsnag/Bugsnag.h>
 
-@interface Bugsnag()
-+ (id) notifier;
-@end
-
 @interface Scenario : NSObject
 
 @property (strong, nonatomic, nonnull) BugsnagConfiguration *config;
@@ -25,7 +21,5 @@
 - (void)run;
 
 - (void)startBugsnag;
-
-- (void)flushAllSessions;
 
 @end